Separate function to get 2000 and 2010 data from the CDC as these datasets aren't stored in the same scheme as other years.

Separate function to get 2000 and 2010 data from the CDC as these datasets aren't stored in the same scheme as other years.

get_2000_2010_data(geography, year, state, geometry)

Arguments

geography

Passed from parent function. Only tract accepted.

year

The year of the SVI to download. Can be 2000 or 2010.

state

The name, abbreviation, or FIPS code of the state for which the SVI is being requested. If NULL (the default), requests data for the entire US.

geometry

if FALSE (the default), return the SVI as tibble. If true, downloads the SVI as a shapefile and returns an sf tibble.

Value

A tibble or sf tibble of the SVI and underlying data for the US or specified state.